requirements.txt
click==8.1.7
datasets==2.19.2
h5py==3.11.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
hypothesis==6.97.6
lightning-bolts==0.7.0
mne-bids==0.15.0
nltk==3.8.1
numpy==1.24.4
omegaconf==2.3.0
pandas==2.0.3
pip==24.0
pre-commit==3.7.1
pytest-benchmark==4.0.0
pytest-cov==5.0.0
pytest-env==1.1.3
pytest-mock==3.14.0
pytest-rerunfailures==14.0
pytest-xdist[psutil]==3.5.0
pytest==7.4.4
python-levenshtein==0.12.2
pytorch-lightning==1.8.6
torch==2.3.0
scipy==1.10.1
setuptools==69.5.1
torchaudio==2.3.0
torchmetrics==0.11.4
trio==0.23.2
unidecode==1.3.8
optuna==3.5.0
plotly==5.20.0
git+https://github.com/stepanhruda/camel-kenlm.git@a6e906b2b30497dd999cb9c84c42c5111f8616e0